pyproject.toml
[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ds-mezo"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "torch==2.10.0",
    "triton==3.6.0",
    "vllm==0.17.0",
    "safetensors==0.7.0",
    "transformers==4.56.0",
    "peft==0.15.0",
    "datasets==3.5.0",
    "evaluate==0.4.3",
    "PyYAML==6.0.1",
    "scipy==1.15.0",
    "numpy==2.2.0",
    "optht==0.2.0",
    "trl==0.16.0",
    "accelerate==1.5.0",
]

[project.scripts]
ds-mezo-train = "scripts.train:main"
ds-mezo-pissa = "scripts.prepare_pissa:main"

[tool.setuptools.packages.find]
include = ["ds_mezo*", "scripts*", "eval*"]
